requirements.txt
dvc
pandas
numpy
scikit-learn
matplotlib
pyyaml
joblib
mlflow